fix(table | code-diff): 添加表格双击事件, 将highlight.js移到peerDependencies #55
affects: @bkui-vue/code-diff, @bkui-vue/table
