+++ site/views/code-diff/index.tsx
@@ -23,7 +23,8 @@
 * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
 * IN THE SOFTWARE.
 */
-import { defineComponent, reactive, ref, Suspense } from 'vue';
+import hljs from 'highlight.js';
+import { defineComponent, onMounted, reactive, ref, Suspense } from 'vue';
 
 import Button from '@bkui-vue/button';
 import BkCodeDiff, { LanguagesUnionType, ThemesUnionType } from '@bkui-vue/code-diff';
@@ -31,10 +32,61 @@ import Input from '@bkui-vue/input';
 
 import { NEW_STR, OLD_STR } from './demo';
 
+/**
+ * 动态引入语法包
+ * [string] language
+ */
+function languageDynamicImport(language) {
+  switch (language) {
+    case 'css':
+      return import('highlight.js/lib/languages/css');
+    case 'java':
+      return import('highlight.js/lib/languages/java');
+    case 'javascript':
+      return import('highlight.js/lib/languages/javascript');
+    case 'json':
+      return import('highlight.js/lib/languages/json');
+    case 'scss':
+      return import('highlight.js/lib/languages/scss');
+    case 'less':
+      return import('highlight.js/lib/languages/less');
+    case 'stylus':
+      return import('highlight.js/lib/languages/stylus');
+    case 'shell':
+      return import('highlight.js/lib/languages/shell');
+    case 'bash':
+      return import('highlight.js/lib/languages/bash');
+    case 'cpp':
+      return import('highlight.js/lib/languages/cpp');
+    case 'go':
+      return import('highlight.js/lib/languages/go');
+    case 'xml':
+      return import('highlight.js/lib/languages/xml');
+    case 'python':
+      return import('highlight.js/lib/languages/python');
+    case 'typescript':
+      return import('highlight.js/lib/languages/typescript');
+    case 'sql':
+      return import('highlight.js/lib/languages/sql');
+    case 'ruby':
+      return import('highlight.js/lib/languages/ruby');
+    case 'vim':
+      return import('highlight.js/lib/languages/vim');
+    case 'php':
+      return import('highlight.js/lib/languages/php');
+    case 'perl':
+      return import('highlight.js/lib/languages/perl');
+    case 'powershell':
+      return import('highlight.js/lib/languages/powershell');
+    case 'makefile':
+      return import('highlight.js/lib/languages/makefile');
+  }
+}
 
 export default defineComponent({
   setup() {
     const theme = ref<ThemesUnionType>('dark');
+    const languageModuleMap = ref({});
     const state = reactive({
       language: 'javascript',
       diffContext: 20,
@@ -48,6 +100,19 @@ export default defineComponent({
     function handleLanChange(val) {
       state.language = val;
     }
+    onMounted(async () => {
+      let languageModule = languageModuleMap.value[state.language];
+      if (languageModule === undefined) {
+        const mod = await languageDynamicImport(state.language);
+        languageModule = mod.default;
+
+        languageModuleMap.value = {
+          ...languageModuleMap.value,
+          language: languageModule,
+        };
+      }
+      hljs.registerLanguage(state.language, languageModule);
+    });
     return () => (
       <div style="margin: 20px">
           <Button onClick={handleClick}>Switch Theme</Button>
@@ -62,14 +127,14 @@ export default defineComponent({
           <Suspense>
             {{
               default: () => (
-                <BkCodeDiff language={state.language as LanguagesUnionType} diffContext={state.diffContext} diffFormat='side-by-side' theme={theme.value} oldContent={OLD_STR} newContent={NEW_STR} />
+                <BkCodeDiff hljs={hljs} language={state.language as LanguagesUnionType} diffContext={state.diffContext} diffFormat='side-by-side' theme={theme.value} oldContent={OLD_STR} newContent={NEW_STR} />
               ),
               fallback: () => (
                 <span>loading</span>
               ),
             }}
           </Suspense>
-          <BkCodeDiff diffFormat='line-by-line' language='javascript' theme='dark' oldContent={OLD_STR} newContent={NEW_STR} />
+          <BkCodeDiff hljs={hljs} diffFormat='line-by-line' language='javascript' theme='dark' oldContent={OLD_STR} newContent={NEW_STR} />
       </div>
     );
   },
